Refactoring code to use new Error type
