+++ crates/cli/cli/src/commands/query/channel/end.rs
@@ -5,7 +5,6 @@ use hermes_cli_framework::output::Output;
 
 use hermes_cosmos_client_components::traits::chain_handle::HasBlockingChainHandle;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
-use hermes_cosmos_relayer::types::error::BaseError;
 
 use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::chain::requests::{IncludeProof, QueryChannelRequest, QueryHeight};
@@ -61,7 +60,7 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnd {
 
         let query_height = if let Some(height) = height {
             let specified_height = Height::new(chain.chain_id.version(), height)
-                .map_err(|e| BaseError::generic(eyre!("Failed to create Height with revision number `{}` and revision height `{height}`. Error: {e}", chain.chain_id.version())))?;
+                .map_err(|e| eyre!("Failed to create Height with revision number `{}` and revision height `{height}`. Error: {e}", chain.chain_id.version()))?;
 
             QueryHeight::Specific(specified_height)
         } else {
@@ -79,7 +78,7 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnd {
                     IncludeProof::No,
                 ) {
                     Ok((channel_end, _)) => Ok(channel_end),
-                    Err(e) => Err(BaseError::relayer(e).into()),
+                    Err(e) => Err(e.into()),
                 }
             })
             .await;
@@ -87,11 +86,11 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnd {
         match channel_end {
             Ok(channel_end) => {
                 if channel_end.state_matches(&State::Uninitialized) {
-                    Err(BaseError::generic(eyre!(
+                    Err(eyre!(
                         "port '{}' & channel '{}' do not exist",
                         self.port_id,
                         self.channel_id,
-                    ))
+                    )
                     .into())
                 } else {
                     Ok(Output::success(channel_end))
