+++ crates/cli/cli/src/commands/query/channel/ends.rs
@@ -6,7 +6,6 @@ use hermes_cli_framework::output::Output;
 
 use hermes_cosmos_client_components::traits::chain_handle::HasBlockingChainHandle;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
-use hermes_cosmos_relayer::types::error::BaseError;
 
 use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::chain::requests::{
@@ -82,7 +81,7 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnds {
 
         let query_height = if let Some(height) = height {
             let specified_height = Height::new(chain_id.version(), height)
-                .map_err(|e| BaseError::generic(eyre!("Failed to create Height with revision number `{}` and revision height `{height}`: {e}", chain_id.version())))?;
+                .map_err(|e| eyre!("Failed to create Height with revision number `{}` and revision height `{height}`: {e}", chain_id.version()))?;
 
             QueryHeight::Specific(specified_height)
         } else {
@@ -100,22 +99,22 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnds {
                         },
                         IncludeProof::No,
                     ) else {
-                        return Err(BaseError::generic(eyre!(
+                        return Err(eyre!(
                             "failed to query channel end for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?}"
-                        )).into());
+                        ).into());
                 };
 
                 if channel_end.state_matches(&State::Uninitialized) {
-                    return Err(BaseError::generic(eyre!(
+                    return Err(eyre!(
                         "{port_id}/{channel_id} on chain {chain_id} @ {query_height:?} is uninitialized",
-                    ))
+                    )
                     .into());
                 }
 
                 let Some(connection_id) = channel_end.connection_hops.first() else {
-                    return Err(BaseError::generic(eyre!(
+                    return Err(eyre!(
                         "missing connection hops for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?}",
-                    )).into());
+                    ).into());
                 };
 
                 let Ok((connection_end, _)) = chain_handle
@@ -126,9 +125,9 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnds {
                         },
                         IncludeProof::No,
                     ) else {
-                        return Err(BaseError::generic(eyre!(
+                        return Err(eyre!(
                             "failed to query connection end for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?}"
-                        )).into());
+                        ).into());
                 };
 
                 let client_id = connection_end.client_id().clone();
@@ -141,9 +140,9 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnds {
                         },
                         IncludeProof::No,
                     ) else {
-                        return Err(BaseError::generic(eyre!(
+                        return Err(eyre!(
                             "failed to query client state for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?}"
-                        )).into());
+                        ).into());
                 };
 
                 let channel_counterparty = channel_end.counterparty().clone();
@@ -151,17 +150,17 @@ impl CommandRunner<CosmosBuilder> for QueryChannelEnds {
                 let counterparty_client_id = connection_counterparty.client_id().clone();
 
                 let Some(counterparty_connection_id) = connection_counterparty.connection_id else {
-                    return Err(BaseError::generic(eyre!(
+                    return Err(eyre!(
                         "connection end for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?} does not have counterparty connection id {connection_end:?}",
-                    )).into());
+                    ).into());
                 };
 
                 let counterparty_port_id = channel_counterparty.port_id().clone();
 
                 let Some(counterparty_channel_id) = channel_counterparty.channel_id else {
-                    return Err(BaseError::generic(eyre!(
+                    return Err(eyre!(
                         "channel end for {port_id}/{channel_id} on chain {chain_id} @ {query_height:?} does not have counterparty channel id {channel_end:?}",
-                    )).into());
+                    ).into());
                 };
 
                 let counterparty_chain_id = client_state.chain_id();
