+++ crates/cli/cli/src/commands/query/clients.rs
@@ -1,17 +1,16 @@
-use std::error::Error as StdError;
 use std::fmt;
 
 use cgp_core::HasErrorType;
 use hermes_cli_components::any_client::contexts::any_counterparty::AnyCounterparty;
 use hermes_cli_components::any_client::types::client_state::AnyClientState;
-use oneline_eyre::eyre::Context;
+use hermes_cosmos_relayer::types::error::Error;
 use tracing::info;
 
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::{json, Output};
 use hermes_cosmos_client_components::types::tendermint::TendermintClientState;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
-use hermes_cosmos_relayer::types::error::BaseError;
+use hermes_cosmos_relayer::types::error::ErrorWrapper;
 use hermes_relayer_components::chain::traits::queries::client_state::CanQueryAllClientStatesWithLatestHeight;
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainIdType;
 use hermes_relayer_components::chain::traits::types::client_state::{
@@ -142,14 +141,13 @@ async fn query_all_client_states<Chain, Counterparty>(
 where
     Chain: HasIbcChainTypes<Counterparty, ClientId = ClientId>
         + CanQueryAllClientStatesWithLatestHeight<Counterparty>
-        + HasErrorType,
+        + HasErrorType<Error = Error>,
     Counterparty: HasClientStateType<Chain, ClientState = AnyClientState>,
-    Chain::Error: From<BaseError> + StdError,
 {
     let mut clients = chain
         .query_all_client_states_with_latest_height()
         .await
-        .wrap_err("Failed to query clients")?
+        .wrap_error("Failed to query clients")?
         .into_iter()
         .map(|(client_id, client_state)| Client::<Chain, Counterparty> {
             client_id,
