+++ crates/cli/cli/src/commands/query/connection/end.rs
@@ -4,7 +4,6 @@ use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
-use hermes_cosmos_relayer::types::error::BaseError as RelayerError;
 use hermes_relayer_components::chain::traits::queries::connection_end::CanQueryConnectionEnd;
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainId;
 use ibc_relayer_types::core::ics03_connection::connection::State;
@@ -48,7 +47,7 @@ impl CommandRunner<CosmosBuilder> for QueryConnectionEnd {
         let chain = builder.build_chain(&self.chain_id).await?;
         let height = self.height.map(|h|
             Height::new(chain.chain_id().version(), h)
-                .map_err(|e| RelayerError::generic(eyre!("Failed to create Height with revision number `{}` and revision height `{h}`. Error: {e}", chain.chain_id().version())))
+                .map_err(|e| eyre!("Failed to create Height with revision number `{}` and revision height `{h}`. Error: {e}", chain.chain_id().version()))
         ).transpose()?;
 
         let connection_end =
