+++ crates/cli/cli/src/commands/query/connections.rs
@@ -1,3 +1,4 @@
+use hermes_cosmos_relayer::types::error::ErrorWrapper;
 use oneline_eyre::eyre::Context;
 use tracing::{info, warn};
 
@@ -7,7 +8,6 @@ use hermes_cli_framework::output::Output;
 use hermes_cosmos_client_components::traits::chain_handle::HasBlockingChainHandle;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
-use hermes_cosmos_relayer::types::error::BaseError;
 use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientStateWithLatestHeight;
 use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::chain::requests::PageRequest;
@@ -55,10 +55,10 @@ impl CommandRunner<CosmosBuilder> for QueryConnections {
                     .query_connections(QueryConnectionsRequest {
                         pagination: Some(PageRequest::all()),
                     })
-                    .map_err(|e| BaseError::relayer(e).into())
+                    .map_err(From::from)
             })
             .await
-            .wrap_err("Failed to query connections for host chain")?;
+            .wrap_error("Failed to query connections for host chain")?;
 
         info!(
             "Found {} connections on chain `{chain_id}`",
