+++ crates/cosmos/cosmos-relayer/src/impls/error.rs
@@ -1,15 +1,19 @@
 use core::num::ParseIntError;
-use std::string::FromUtf8Error;
 
-use cgp_core::{Async, ErrorRaiser, HasErrorType, ProvideErrorType};
-use eyre::eyre;
+use alloc::string::FromUtf8Error;
+use cgp_core::prelude::*;
+use cgp_core::{ErrorRaiser, ProvideErrorType};
+use eyre::Report;
 use hermes_cli_components::any_client::impls::decoders::client_state::UnknownClientStateType;
 use hermes_cosmos_client_components::impls::decoders::type_url::TypeUrlMismatchError;
 use hermes_cosmos_client_components::impls::queries::abci::AbciQueryError;
+use hermes_relayer_components::chain::traits::types::chain_id::HasChainIdType;
 use hermes_relayer_components::relay::impls::create_client::MissingCreateClientEventError;
 use hermes_relayer_runtime::types::error::TokioRuntimeError;
 use hermes_test_components::chain::impls::assert::poll_assert_eventual_amount::EventualAmountTimeoutError;
 use hermes_test_components::chain::impls::ibc_transfer::MissingSendPacketEventError;
+use hermes_test_components::chain::traits::types::address::HasAddressType;
+use hermes_test_components::chain::traits::types::amount::HasAmountType;
 use ibc_relayer::error::Error as RelayerError;
 use ibc_relayer::supervisor::Error as SupervisorError;
 use ibc_relayer_types::core::ics02_client::error::Error as Ics02Error;
@@ -19,7 +23,9 @@ use tendermint_proto::Error as TendermintProtoError;
 use tendermint_rpc::Error as TendermintRpcError;
 
 use crate::contexts::chain::CosmosChain;
-use crate::types::error::{BaseError, Error};
+use crate::types::error::{
+    DebugNonRetryableError, DisplayNonRetryableError, Error, ReportNonRetryableError, ReturnError,
+};
 
 pub struct HandleCosmosError;
 
@@ -30,189 +36,74 @@ where
     type Error = Error;
 }
 
-impl<Context> ErrorRaiser<Context, Error> for HandleCosmosError
+pub trait CheckErrorRaiser<Context>:
+    ErrorRaiser<Context, TokioRuntimeError>
+    + for<'a> ErrorRaiser<Context, &'a str>
+    + for<'a> ErrorRaiser<Context, EventualAmountTimeoutError<'a, CosmosChain>>
 where
     Context: HasErrorType<Error = Error>,
 {
-    fn raise_error(e: Error) -> Error {
-        e
+}
+
+impl<Context> CheckErrorRaiser<Context> for HandleCosmosError where
+    Context: HasErrorType<Error = Error>
+{
+}
+
+impl<Context, E, Delegate> ErrorRaiser<Context, E> for HandleCosmosError
+where
+    Context: HasErrorType,
+    Self: DelegateComponent<E, Delegate = Delegate>,
+    Delegate: ErrorRaiser<Context, E>,
+{
+    fn raise_error(e: E) -> Context::Error {
+        Delegate::raise_error(e)
     }
 }
 
-impl<Context> ErrorRaiser<Context, TokioRuntimeError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: TokioRuntimeError) -> Error {
-        BaseError::tokio(e).into()
+delegate_components! {
+    HandleCosmosError {
+        Error: ReturnError,
+        [
+            Report,
+            TokioRuntimeError,
+            RelayerError,
+            SupervisorError,
+            TendermintProtoError,
+            TendermintRpcError,
+            Ics02Error,
+            Ics24ValidationError,
+            ParseIntError,
+            FromUtf8Error,
+            EncodeError,
+            DecodeError,
+        ]: ReportNonRetryableError,
+        [
+            TypeUrlMismatchError,
+            UnknownClientStateType,
+            AbciQueryError,
+            MissingSendPacketEventError,
+        ]:
+            DebugNonRetryableError,
     }
 }
 
-impl<Context> ErrorRaiser<Context, RelayerError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(err: RelayerError) -> Error {
-        BaseError::relayer(err).into()
-    }
+impl<'a> DelegateComponent<&'a str> for HandleCosmosError {
+    type Delegate = DisplayNonRetryableError;
 }
 
-impl<Context> ErrorRaiser<Context, SupervisorError> for HandleCosmosError
+impl<'a, Chain> DelegateComponent<EventualAmountTimeoutError<'a, Chain>> for HandleCosmosError
 where
-    Context: HasErrorType<Error = Error>,
+    Chain: HasAddressType + HasAmountType,
 {
-    fn raise_error(err: SupervisorError) -> Error {
-        BaseError::supervisor(err).into()
-    }
+    type Delegate = DebugNonRetryableError;
 }
 
-impl<Context> ErrorRaiser<Context, TendermintProtoError> for HandleCosmosError
+impl<'a, Chain, Counterparty>
+    DelegateComponent<MissingCreateClientEventError<'a, Chain, Counterparty>> for HandleCosmosError
 where
-    Context: HasErrorType<Error = Error>,
+    Chain: HasChainIdType,
+    Counterparty: HasChainIdType,
 {
-    fn raise_error(e: TendermintProtoError) -> Error {
-        BaseError::generic(e.into()).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, AbciQueryError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: AbciQueryError) -> Error {
-        BaseError::generic(eyre!("abci query returned error: {:?}", e.response)).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, TendermintRpcError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(err: TendermintRpcError) -> Error {
-        BaseError::tendermint_rpc(err).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, ParseIntError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: ParseIntError) -> Error {
-        BaseError::generic(e.into()).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, FromUtf8Error> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: FromUtf8Error) -> Error {
-        BaseError::generic(e.into()).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, MissingSendPacketEventError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(_e: MissingSendPacketEventError) -> Error {
-        BaseError::generic(eyre!("missing send packet event")).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, Ics02Error> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(err: Ics02Error) -> Error {
-        BaseError::ics02(err).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, Ics24ValidationError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(err: Ics24ValidationError) -> Error {
-        BaseError::ics24_validation(err).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, TypeUrlMismatchError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: TypeUrlMismatchError) -> Error {
-        BaseError::generic(eyre!(
-            "type url mismatch. expected: {}, actual: {}",
-            e.expected_url,
-            e.actual_url
-        ))
-        .into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, UnknownClientStateType> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: UnknownClientStateType) -> Error {
-        BaseError::generic(eyre!("unknown client state type: {}", e.type_url,)).into()
-    }
-}
-
-impl<'a, Context> ErrorRaiser<Context, EventualAmountTimeoutError<'a, CosmosChain>>
-    for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: EventualAmountTimeoutError<'a, CosmosChain>) -> Error {
-        BaseError::generic(eyre!("{:?}", e)).into()
-    }
-}
-
-impl<'a, Context> ErrorRaiser<Context, MissingCreateClientEventError<'a, CosmosChain, CosmosChain>>
-    for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: MissingCreateClientEventError<'a, CosmosChain, CosmosChain>) -> Error {
-        BaseError::generic(eyre!("{:?}", e)).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, EncodeError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: EncodeError) -> Error {
-        BaseError::generic(e.into()).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, DecodeError> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: DecodeError) -> Error {
-        BaseError::generic(e.into()).into()
-    }
-}
-
-impl<'a, Context> ErrorRaiser<Context, &'a str> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: &'a str) -> Error {
-        BaseError::generic(eyre!("{e}")).into()
-    }
-}
-
-impl<Context> ErrorRaiser<Context, eyre::Report> for HandleCosmosError
-where
-    Context: HasErrorType<Error = Error>,
-{
-    fn raise_error(e: eyre::Report) -> Error {
-        BaseError::generic(e).into()
-    }
+    type Delegate = DebugNonRetryableError;
 }
