+++ crates/cosmos/cosmos-relayer/src/impls/mod.rs
@@ -1,7 +1,6 @@
 pub mod birelay;
 pub mod build;
 pub mod error;
-pub mod error2;
 pub mod relay;
 pub mod subscription;
 pub mod telemetry;
