+++ crates/cosmos/cosmos-relayer/src/impls/transaction/components/estimate.rs
@@ -1,4 +1,5 @@
 use cgp_core::prelude::*;
+use eyre::eyre;
 use hermes_relayer_components::transaction::traits::components::tx_fee_estimater::TxFeeEstimator;
 use ibc_proto::cosmos::tx::v1beta1::{Fee, Tx};
 use ibc_relayer::chain::cosmos::gas::gas_amount_to_fee;
@@ -7,7 +8,7 @@ use ibc_relayer::chain::cosmos::types::tx::SignedTx;
 
 use crate::contexts::transaction::CosmosTxContext;
 use crate::impls::transaction::component::CosmosTxComponents;
-use crate::types::error::{BaseError, Error};
+use crate::types::error::Error;
 
 #[async_trait]
 impl TxFeeEstimator<CosmosTxContext> for CosmosTxComponents {
@@ -20,13 +21,11 @@ impl TxFeeEstimator<CosmosTxContext> for CosmosTxComponents {
 
         let tx_config = &context.tx_config;
 
-        let response = send_tx_simulate(&tx_config.grpc_address, tx)
-            .await
-            .map_err(BaseError::relayer)?;
+        let response = send_tx_simulate(&tx_config.grpc_address, tx).await?;
 
         let gas_info = response
             .gas_info
-            .ok_or_else(BaseError::missing_simulate_gas_info)?;
+            .ok_or_else(|| eyre!("missing simulate gas info"))?;
 
         let fee = gas_amount_to_fee(&tx_config.gas_config, gas_info.gas_used);
