+++ crates/cosmos/cosmos-relayer/src/impls/transaction/error.rs
@@ -1,11 +1,12 @@
+use eyre::eyre;
 use hermes_relayer_components::transaction::components::poll_tx_response::CanRaiseNoTxResponseError;
 use tendermint::Hash as TxHash;
 
 use crate::contexts::transaction::CosmosTxContext;
-use crate::types::error::{BaseError, Error};
+use crate::types::error::Error;
 
 impl CanRaiseNoTxResponseError for CosmosTxContext {
     fn tx_no_response_error(tx_hash: &TxHash) -> Error {
-        BaseError::tx_no_response(*tx_hash).into()
+        eyre!("failed to receive tx response for tx hash: {}", tx_hash).into()
     }
 }
