+++ crates/cosmos/cosmos-relayer/src/types/mod.rs
@@ -1,4 +1,3 @@
 pub mod batch;
 pub mod error;
-pub mod error2;
 pub mod telemetry;
